# Schems/swagger.yaml
openapi: 3.0.4
info:
  title: OpenVPN Gate Monitor API
  version: v1
paths:
...
  '/api/open-vpn-servers/get-server-with-status/{vpnServerId}':
    get:
      tags:
        - OpenVpnServers
      parameters:
        - name: VpnServerId
          in: path
          required: true
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OpenVpnServerWithStatusResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerWithStatusResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerWithStatusResponseApiResponse'
...
  '/api/open-vpn-servers/get/{vpnServerId}':
    get:
      tags:
        - OpenVpnServers
      parameters:
        - name: VpnServerId
          in: path
          required: true
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
  /api/open-vpn-servers/add:
    post:
      tags:
        - OpenVpnServers
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/AddServerRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/AddServerRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AddServerRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AddServerRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
  /api/open-vpn-servers/update:
    put:
      tags:
        - OpenVpnServers
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateServerRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/OpenVpnServerResponseApiResponse'
  '/api/open-vpn-servers/delete/{vpnServerId}':
    delete:
      tags:
        - OpenVpnServers
      parameters:
        - name: VpnServerId
          in: path
          required: true
          schema:
            maximum: 2147483647
            minimum: 1
            type: integer
            format: int32
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    AddServerRequest:
      required:
        - serverName
      type: object
      properties:
        serverName:
          minLength: 1
          type: string
        isOnline:
          type: boolean
        isDefault:
          type: boolean
        apiUrl:
          type: string
          nullable: true
      additionalProperties: false
...
    OpenVpnServerDto:
      type: object
      properties:
        id:
          type: integer
          format: int32
        serverName:
          type: string
          nullable: true
        isOnline:
          type: boolean
        isDefault:
          type: boolean
        apiUrl:
          type: string
          nullable: true
        createDate:
          type: string
          format: date-time
        lastUpdate:
          type: string
          format: date-time
      additionalProperties: false
...
    OpenVpnServerResponse:
      type: object
      properties:
        openVpnServer:
          $ref: '#/components/schemas/OpenVpnServerDto'
      additionalProperties: false
    OpenVpnServerResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OpenVpnServerResponse'
      additionalProperties: false
    OpenVpnServerStatusLogResponse:
      type: object
      properties:
        vpnServerId:
          type: integer
          format: int32
        sessionId:
          type: string
          format: uuid
        upSince:
          type: string
          format: date-time
        serverLocalIp:
          type: string
          nullable: true
        serverRemoteIp:
          type: string
          nullable: true
        bytesIn:
          type: integer
          format: int64
        bytesOut:
          type: integer
          format: int64
        version:
          type: string
          nullable: true
      additionalProperties: false
    OpenVpnServerWithStatusDto:
      type: object
      properties:
        openVpnServerResponses:
          $ref: '#/components/schemas/OpenVpnServerResponse'
        openVpnServerStatusLogResponse:
          $ref: '#/components/schemas/OpenVpnServerStatusLogResponse'
        countConnectedClients:
          type: integer
          format: int32
        countSessions:
          type: integer
          format: int32
        totalBytesIn:
          type: integer
          format: int64
        totalBytesOut:
          type: integer
          format: int64
      additionalProperties: false
    OpenVpnServerWithStatusResponse:
      type: object
      properties:
        openVpnServerWithStatus:
          $ref: '#/components/schemas/OpenVpnServerWithStatusDto'
      additionalProperties: false
    OpenVpnServerWithStatusResponseApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
          nullable: true
        data:
          $ref: '#/components/schemas/OpenVpnServerWithStatusResponse'
      additionalProperties: false
...
    UpdateServerRequest:
      required:
        - id
        - serverName
      type: object
      properties:
        id:
          maximum: 2147483647
          minimum: 1
          type: integer
          format: int32
        serverName:
          minLength: 1
          type: string
        isOnline:
          type: boolean
        isDefault:
          type: boolean
        apiUrl:
          type: string
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: http
      description: 'Enter the token in the format ''Bearer {your_token}'''
      scheme: Bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]